src/frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "AI Applications Chassis - AI Apps Best Practices Tailored for Azure - Frontend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv ~=1.0.1",
    "semantic-kernel>=1.32.2",
    "rich>=14.0.0",
    "chainlit>=2.5.5",
    "azure-monitor-opentelemetry-exporter>=1.0.0b38",
    "opentelemetry-exporter-otlp-proto-grpc>=1.31.1",
    "opentelemetry-instrumentation>=0.52b1",
    "azure-monitor-opentelemetry>=1.6.10",
    "opentelemetry-sdk>=1.31.1",
    "azure-ai-projects>=1.0.0b11",
    "opentelemetry-instrumentation-openai-v2>=2.1b0",
    "azure-core-tracing-opentelemetry>=1.0.0b12",
]
